Moved trait implementation to separate test.
