+++ crates/lang/macro/tests/ui/pass/09-static-env.rs
