+++ crates/lang/macro/tests/ui/pass/10-derive-for-storage.rs
