+++ crates/lang/macro/tests/ui/pass/11-alias-storage-struct-impl.rs
